pyproject.toml
[project]
name = "kl3m-data"
version = "0.1.1"
description = "KL3M training data collection and preprocesing from the ALEA Institute"
authors = [{ name = "ALEA Institute", email = "hello@aleainstitute.ai" }]
requires-python = ">=3.10,<4.0.0"
readme = "README.md"
license = "MIT"
keywords = [
    "alea",
    "kl3m",
    "data",
    "training data",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Utilities",
]
dependencies = [
    "alea-dublincore>=0.1.0,<0.2",
    "rich>=13.8.1,<14",
    "httpx[http2, socks]>=0.27.2,<0.28",
    "playwright>=1.47.0,<2",
    "lxml>=5.3.0,<6",
    "boto3>=1.35.22,<2",
    "pypdfium2>=4.30.0,<5",
    "alea-preprocess>=0.1.10,<0.2",
    "tqdm>=4.66.5,<5",
    "pyyaml>=6.0.2",
    "pyarrow>=19.0.0",
    "tokenizers>=0.21.0",
    "datasets>=3.2.0",
]

[project.urls]
Homepage = "https://aleainstitute.ai/"
Repository = "https://github.com/alea-institute/kl3m-data"

[dependency-groups]
dev = [
    "pytest>=8.3.2,<9",
    "pytest-asyncio>=0.23.8,<0.24",
    "pytest-benchmark>=4.0.0,<5",
    "pytest-cov>=5.0.0,<6",
    "pylint>=3.2.7,<4",
    "mypy>=1.11.2,<2",
]

[tool.hatch.build.targets.sdist]
include = [
    "kl3m_data",
    "kl3m_data",
]
exclude = [
    "tests",
    "docs",
    "examples",
    "docker",
]

[tool.hatch.build.targets.wheel]
include = [
    "kl3m_data",
    "kl3m_data",
]
exclude = [
    "tests",
    "docs",
    "examples",
    "docker",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pylint]
max-line-length = 120
disable = [
    "line-too-long",
    "too-few-public-methods",
    "too-many-positional-arguments",
    "no-self-argument",
    "cyclic-import",
    "duplicate-code",
]
extension-pkg-allow-list = ["lxml"]
ignore-patterns = ["test_.*?.py", "conftest.py"]

[tool.pylint.messages_control]
max-args = 10
max-attributes = 15
max-branches = 15
max-locals = 25
min-public-methods = 0
disable = ['W0511']

[tool.isort]
profile = "black"
line_length = 120

[tool.pytest.ini_options]
addopts = "--cov=kl3m_data --cov-report=term-missing --cov-report=xml --cov-report=html"

[tool.mypy]
ignore_missing_imports = true
